git comente lo que se dijo en la daily
